+++ History.md
@@ -3,6 +3,7 @@
 
 * Improved performance of advanced mode validators.
 * Fixed issue [#307](https://github.com/GoalSmashers/clean-css/issues/307) - background-color in muliple backgrounds.
+* Fixed issue [#322](https://github.com/GoalSmashers/clean-css/issues/322) - adds background-size support.
 * Fixed issue [#323](https://github.com/GoalSmashers/clean-css/issues/323) - stripping variable references.
 * Fixed issue [#325](https://github.com/GoalSmashers/clean-css/issues/325) - removing invalid @charset declarations.